Break header into components
