+++ public/global.css
@@ -11,8 +11,8 @@ body {
   margin: 0;
   padding: 8px;
   box-sizing: border-box;
-  font-family: "Inter", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans,
-    Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
+  font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen-Sans, Ubuntu, Cantarell,
+    'Helvetica Neue', sans-serif;
 }
 
 a {
@@ -21,7 +21,6 @@ a {
 }
 
 a:hover {
-  text-decoration: underline;
 }
 
 a:visited {
@@ -182,7 +181,7 @@ textarea {
   border-radius: 2px;
   background: #9c9ea1;
   cursor: default;
-  transition: all .25s ease-out;
+  transition: all 0.25s ease-out;
   cursor: pointer;
 }
