+++ src/components/SocialButtons.svelte
@@ -1,0 +1,48 @@
+<div class="flex flex-row">
+  <a href="https://ko-fi.com/T6T31IWGF" target="_blank">
+    <button
+      class="text-sm rounded bg-gray-700 transition-all duration-300 ease-out hover:bg-gray-500 px-4 py-2
+      whitespace-no-wrap flex flex-row items-center justify-center">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        class="feather feather-coffee mr-2">
+        <path d="M18 8h1a4 4 0 0 1 0 8h-1" />
+        <path d="M2 8h16v9a4 4 0 0 1-4 4H6a4 4 0 0 1-4-4V8z" />
+        <line x1="6" y1="1" x2="6" y2="4" />
+        <line x1="10" y1="1" x2="10" y2="4" />
+        <line x1="14" y1="1" x2="14" y2="4" />
+      </svg>
+      Buy me a coffee
+    </button>
+  </a>
+  <a href="https://twitter.com/imjoshellis">
+    <button
+      class="rounded bg-gray-800 text-gray-500 transition-all duration-300 ease-out hover:text-gray-300
+      hover:bg-blue-600 ml-2 px-4 py-2 text-sm flex flex-row items-center justify-center">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        class="feather feather-twitter mr-2">
+        <path
+          d="M23 3a10.9 10.9 0 0 1-3.14 1.53 4.48 4.48 0 0 0-7.86 3v1A10.66 10.66 0 0 1 3 4s-4 9 5 13a11.64 11.64 0 0
+          1-7 2c9 5 20 0 20-11.5a4.5 4.5 0 0 0-.08-.83A7.72 7.72 0 0 0 23 3z" />
+      </svg>
+      @imjoshellis
+    </button>
+  </a>
+</div>
